Combine start and reuse order fees to remove duplication in asset creation and start order.
